@@ -44,8 +44,9 @@
     "@map-colonies/mc-priority-queue": "^7.1.0",
     "@map-colonies/read-pkg": "0.0.1",
     "@map-colonies/telemetry": "^6.0.0",
-    "@opentelemetry/api": "^1.7.1",
+    "@opentelemetry/api": "1.7.0",
     "@opentelemetry/api-metrics": "0.23.0",
+    "@opentelemetry/context-async-hooks": "^1.24.1",
     "@opentelemetry/instrumentation-express": "0.32.1",
     "@opentelemetry/instrumentation-http": "0.35.1",
     "axios": "^1.4.0",
@@ -78,6 +79,6 @@
     "rimraf": "^5.0.1",
     "standard-version": "^9.5.0",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.1.3"
+    "typescript": "^5.3.3"
   }
 }
